backend/requirements.txt
aniso8601==9.0.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cloudinary==1.25.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.1
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
six==1.16.0
twilio==6.59.0
urllib3==1.26.4
Werkzeug==2.0.1
